+++ usipipo_commons/domain/entities/__init__.py
@@ -2,6 +2,7 @@ from .consumption_billing import ConsumptionBilling
 from .consumption_invoice import ConsumptionInvoice
 from .crypto_order import CryptoOrder
 from .payment import Payment
+from .subscription_plan import SubscriptionPlan
 from .user import User
 from .vpn_key import VpnKey
 
@@ -12,4 +13,5 @@ __all__ = [
     "ConsumptionBilling",
     "ConsumptionInvoice",
     "CryptoOrder",
+    "SubscriptionPlan",
 ]